+++ docs/examples/Image_encodeToData.cpp
@@ -8,7 +8,9 @@ void draw(SkCanvas* canvas) {
     SkIRect subset = {0, 0, 16, 64};
     int x = 0;
     for (int quality : { 0, 10, 50, 100 } ) {
-        sk_sp<SkData> data(image->encodeToData(SkEncodedImageFormat::kJPEG, quality));
+        SkJpegEncoder::Options options;
+        options.fQuality = quality;
+        sk_sp<SkData> data(SkJpegEncoder::Encode(nullptr, image.get(), options));
         sk_sp<SkImage> filtered = SkImages::DeferredFromEncodedData(data)->makeSubset(subset);
         canvas->drawImage(filtered, x, 0);
         x += 16;