+++ gm/image_shader.cpp
@@ -26,6 +26,7 @@
 #include "include/core/SkSurface.h"
 #include "include/core/SkTileMode.h"
 #include "include/core/SkTypes.h"
+#include "include/encode/SkPngEncoder.h"
 #include "include/gpu/GpuTypes.h"
 
 #include <utility>
@@ -86,7 +87,7 @@ static sk_sp<SkImage> make_encode_gen(GrRecordingContext* ctx,
     if (!src) {
         return nullptr;
     }
-    sk_sp<SkData> encoded = src->encodeToData(SkEncodedImageFormat::kPNG, 100);
+    sk_sp<SkData> encoded = SkPngEncoder::Encode(nullptr, src.get(), {});
     if (!encoded) {
         return nullptr;
     }