+++ src/image/SkImage.cpp
@@ -199,6 +199,7 @@ sk_sp<SkShader> SkImage::makeRawShader(SkTileMode tmx, SkTileMode tmy,
                                   sampling, localMatrix);
 }
 
+#if !defined(SK_DISABLE_LEGACY_IMAGE_ENCODE_METHODS)
 sk_sp<SkData> SkImage::encodeToData(GrDirectContext* context, SkEncodedImageFormat type,
                                     int quality) const {
     SkBitmap bm;
@@ -216,7 +217,6 @@ sk_sp<SkData> SkImage::encodeToData(GrDirectContext* context) const {
     return this->encodeToData(context, SkEncodedImageFormat::kPNG, 100);
 }
 
-#ifndef SK_IMAGE_READ_PIXELS_DISABLE_LEGACY_API
 sk_sp<SkData> SkImage::encodeToData(SkEncodedImageFormat type, int quality) const {
     auto dContext = as_IB(this)->directContext();
     return this->encodeToData(dContext, type, quality);